+++ src/main/java/edu/jojos/bizarre/matrix/paging/reference/PageReference.java
@@ -1,3 +1,3 @@
 package edu.jojos.bizarre.matrix.paging.reference;
 
-public record PageReference(String metadata, long pageNumber, long pageOffset) {}
+public record PageReference(String metadata, int pageNumber, int pageOffset) {}
